third_party/WebKit/LayoutTests/http/tests/htmlimports/import-dynamic-async-notblock-sync.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
(function() {
    var link = document.createElement('link');
    link.id = 'dynamicLink'
    link.rel = 'import';
    link.setAttribute('async', '');
    link.href = 'resources/import-slow-child.cgi';
    document.head.appendChild(link);
})();
</script>
<link id="followingLink" rel="import" href="resources/hello.html">
<script>
test(function() {
   assert_true(null == dynamicLink.import, 'Dynamic import'); 
   assert_true(null != followingLink.import, 'Following import'); 
}, 'Dynamically added async import doesn\'t block following import.');
</script>
</body>
</html>
